refactor: 비즈니스 로직을 가진 Class, domain 패키지로 이동
