+++ src/main/java/racingcar/domain/AlwaysMoveCarMover.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class AlwaysMoveCarMover implements CarMover {
