+++ src/main/java/racingcar/domain/CarMover.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface CarMover {
     public boolean movable();
