+++ src/main/java/racingcar/domain/RacingVehicle.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 /**
  * @author daheekim
