+++ src/main/java/racingcar/domain/RepeatingCarMover.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class RepeatingCarMover implements CarMover {
     private static boolean flag = true;
