+++ src/main/java/racingcar/view/InputView.java
@@ -1,6 +1,5 @@
 package racingcar.view;
 
-import racingcar.RacingGame;
 import util.StringUtils;
 
 import java.util.Arrays;
